requirements.txt
asgiref==3.2.3
cffi==1.13.2
chardet==3.0.4
Cython==0.29.14
Django==3.0.2
img2pdf==0.3.3
lxml==4.4.2
ocrmypdf==9.4.0
pdfminer.six==20200104
pikepdf==1.8.3
Pillow==7.0.0
pycparser==2.19
pycryptodome==3.9.4
python-magic==0.4.15
pytz==2019.3
reportlab==3.5.34
sortedcontainers==2.1.0
sqlparse==0.3.0
tqdm==4.41.1
